+++ packages/app/src/cli/commands/app/dev.ts
@@ -4,7 +4,7 @@ import dev from '$cli/services/dev'
 import {load as loadApp, App} from '$cli/models/app/app'
 
 export default class Dev extends Command {
-  static description = 'Develop a block or an app'
+  static description = 'Run the app'
 
   static flags = {
     path: Flags.string({