+++ packages/app/src/cli/commands/app/scaffold/extension.ts
@@ -1,26 +1,26 @@
 import {output, path} from '@shopify/cli-kit'
 import {Command, Flags} from '@oclif/core'
-import {uiExtensions} from '$cli/constants'
-import scaffoldUiExtensionPrompt from '$cli/prompts/scaffold/ui-extension'
+import {extensions} from '$cli/constants'
+import scaffoldExtensionPrompt from '$cli/prompts/scaffold/extension'
 import {load as loadApp, App} from '$cli/models/app/app'
-import scaffoldUiExtensionService from '$cli/services/scaffold/ui-extension'
+import scaffoldExtensionService from '$cli/services/scaffold/extension'
 
-export default class AppScaffoldUiExtension extends Command {
-  static description = 'Scaffold a UI Extension'
+export default class AppScaffoldExtension extends Command {
+  static description = 'Scaffold an Extension'
   static examples = ['<%= config.bin %> <%= command.id %>']
 
   static flags = {
     type: Flags.string({
       char: 't',
       hidden: false,
-      description: 'UI Extension type',
-      options: uiExtensions.types,
+      description: 'Extension type',
+      options: extensions.types,
       env: 'SHOPIFY_FLAG_TYPE',
     }),
     name: Flags.string({
       char: 'n',
       hidden: false,
-      description: 'name of your UI Extension',
+      description: 'name of your Extension',
       env: 'SHOPIFY_FLAG_NAME',
     }),
     path: Flags.string({
@@ -34,18 +34,18 @@ export default class AppScaffoldUiExtension extends Command {
   static args = [{name: 'file'}]
 
   public async run(): Promise<void> {
-    const {args, flags} = await this.parse(AppScaffoldUiExtension)
+    const {args, flags} = await this.parse(AppScaffoldExtension)
     const directory = flags.path ? path.resolve(flags.path) : process.cwd()
     const parentApp: App = await loadApp(directory)
-    const promptAnswers = await scaffoldUiExtensionPrompt({
-      uiExtensionType: flags.type,
+    const promptAnswers = await scaffoldExtensionPrompt({
+      extensionType: flags.type,
       name: flags.name,
     })
-    const {uiExtensionType, name} = promptAnswers
-    await scaffoldUiExtensionService({
+    const {extensionType, name} = promptAnswers
+    await scaffoldExtensionService({
       ...promptAnswers,
       parentApp,
     })
-    output.info(output.content`UI Extension ${name} generated successfully!`)
+    output.info(output.content`Extension ${name} generated successfully!`)
   }
 }