+++ packages/app/src/cli/models/app/app.test.ts
@@ -1,10 +1,10 @@
-import {load, HomeNotFoundError} from './app'
+import {load} from './app'
 import {describe, it, expect, beforeEach, afterEach} from 'vitest'
 import {file, path} from '@shopify/cli-kit'
 import {configurationFileNames, blocks, genericConfigurationFileNames} from '$cli/constants'
 
 describe('load', () => {
-  type BlockType = 'uiExtensions' | 'scripts'
+  type BlockType = 'extensions' | 'scripts'
 
   let tmpDir: string
   const appConfiguration = `
@@ -21,12 +21,16 @@ name = "my_app"
   })
 
   const writeConfig = async (appConfiguration: string) => {
-    const appConfigurationPath = path.join(tmpDir, configurationFileNames.app)
-    await file.write(appConfigurationPath, appConfiguration)
-  }
-
-  const mkdirHome = async () => {
-    await file.mkdir(path.join(tmpDir, 'home'))
+    const appDirectory = path.join(tmpDir, configurationFileNames.app)
+    const homeDirectory = path.join(tmpDir, blocks.home.directoryName)
+    const homeConfiguration = `
+    [commands]
+    build = "build"
+    dev = "dev"
+    `
+    await file.write(appDirectory, appConfiguration)
+    await file.mkdir(homeDirectory)
+    await file.write(path.join(homeDirectory, blocks.home.configurationName), homeConfiguration)
   }
 
   const blockConfigurationPath = ({blockType, name}: {blockType: BlockType; name: string}) => {
@@ -74,7 +78,6 @@ name = "my_app"
         wrong = "my_app"
         `
     await writeConfig(appConfiguration)
-    await mkdirHome()
 
     // When/Then
     await expect(load(tmpDir)).rejects.toThrow()
@@ -83,7 +86,6 @@ name = "my_app"
   it('loads the app when the configuration is valid and has no blocks', async () => {
     // Given
     await writeConfig(appConfiguration)
-    await mkdirHome()
 
     // When
     const app = await load(tmpDir)
@@ -95,7 +97,6 @@ name = "my_app"
   it('defaults to npm as package manager when the configuration is valid', async () => {
     // Given
     await writeConfig(appConfiguration)
-    await mkdirHome()
 
     // When
     const app = await load(tmpDir)
@@ -107,7 +108,6 @@ name = "my_app"
   it('defaults to yarn st the package manager when yarn.lock is present, the configuration is valid, and has no blocks', async () => {
     // Given
     await writeConfig(appConfiguration)
-    await mkdirHome()
     const yarnLockPath = path.join(tmpDir, genericConfigurationFileNames.yarn.lockfile)
     await file.write(yarnLockPath, '')
 
@@ -121,7 +121,6 @@ name = "my_app"
   it('defaults to pnpm st the package manager when pnpm lockfile is present, the configuration is valid, and has no blocks', async () => {
     // Given
     await writeConfig(appConfiguration)
-    await mkdirHome()
     const pnpmLockPath = path.join(tmpDir, genericConfigurationFileNames.pnpm.lockfile)
     await file.write(pnpmLockPath, '')
 
@@ -132,21 +131,21 @@ name = "my_app"
     expect(app.packageManager).toBe('pnpm')
   })
 
-  it("throws an error if the UI extension configuration file doesn't exist", async () => {
+  it("throws an error if the extension configuration file doesn't exist", async () => {
     // Given
-    makeBlockDir({blockType: 'uiExtensions', name: 'my-extension'})
+    makeBlockDir({blockType: 'extensions', name: 'my-extension'})
 
     // When
     await expect(load(tmpDir)).rejects.toThrow(/Couldn't find the configuration file/)
   })
 
-  it('throws an error if the UI extension configuration file is invalid', async () => {
+  it('throws an error if the extension configuration file is invalid', async () => {
     // Given
     const blockConfiguration = `
       wrong = "my_extension"
       `
     await writeBlockConfig({
-      blockType: 'uiExtensions',
+      blockType: 'extensions',
       blockConfiguration,
       name: 'my-extension',
     })
@@ -155,16 +154,15 @@ name = "my_app"
     await expect(load(tmpDir)).rejects.toThrow()
   })
 
-  it('loads the app when it has a UI extension with a valid configuration', async () => {
+  it('loads the app when it has a extension with a valid configuration', async () => {
     // Given
     await writeConfig(appConfiguration)
-    await mkdirHome()
     const blockConfiguration = `
       name = "my_extension"
       type = "checkout-post-purchase"
       `
     await writeBlockConfig({
-      blockType: 'uiExtensions',
+      blockType: 'extensions',
       blockConfiguration,
       name: 'my-extension',
     })
@@ -173,19 +171,18 @@ name = "my_app"
     const app = await load(tmpDir)
 
     // Then
-    expect(app.uiExtensions[0].configuration.name).toBe('my_extension')
+    expect(app.extensions[0].configuration.name).toBe('my_extension')
   })
 
-  it('loads the app from a UI extension directory when it has a UI extension with a valid configuration', async () => {
+  it('loads the app from a extension directory when it has a extension with a valid configuration', async () => {
     // Given
     await writeConfig(appConfiguration)
-    await mkdirHome()
     const blockConfiguration = `
       name = "my_extension"
       type = "checkout-post-purchase"
       `
     const {blockDir} = await writeBlockConfig({
-      blockType: 'uiExtensions',
+      blockType: 'extensions',
       blockConfiguration,
       name: 'my-extension',
     })
@@ -195,20 +192,19 @@ name = "my_app"
 
     // Then
     expect(app.configuration.name).toBe('my_app')
-    expect(app.uiExtensions[0].configuration.name).toBe('my_extension')
+    expect(app.extensions[0].configuration.name).toBe('my_extension')
   })
 
-  it('loads the app with several UI extensions that have valid configurations', async () => {
+  it('loads the app with several extensions that have valid configurations', async () => {
     // Given
     await writeConfig(appConfiguration)
-    await mkdirHome()
 
     let blockConfiguration = `
       name = "my_extension_1"
       type = "checkout-post-purchase"
       `
     await writeBlockConfig({
-      blockType: 'uiExtensions',
+      blockType: 'extensions',
       blockConfiguration,
       name: 'my_extension_1',
     })
@@ -218,7 +214,7 @@ name = "my_app"
       type = "product-subscription"
       `
     await writeBlockConfig({
-      blockType: 'uiExtensions',
+      blockType: 'extensions',
       blockConfiguration,
       name: 'my_extension_2',
     })
@@ -227,9 +223,9 @@ name = "my_app"
     const app = await load(tmpDir)
 
     // Then
-    expect(app.uiExtensions).toHaveLength(2)
-    expect(app.uiExtensions[0].configuration.name).toBe('my_extension_1')
-    expect(app.uiExtensions[1].configuration.name).toBe('my_extension_2')
+    expect(app.extensions).toHaveLength(2)
+    expect(app.extensions[0].configuration.name).toBe('my_extension_1')
+    expect(app.extensions[1].configuration.name).toBe('my_extension_2')
   })
 
   it("throws an error if the configuration file doesn't exist", async () => {
@@ -255,18 +251,9 @@ name = "my_app"
     await expect(load(tmpDir)).rejects.toThrowError()
   })
 
-  it('throws an error if the home directory is missing', async () => {
-    // Given
-    await writeConfig(appConfiguration)
-
-    // When
-    await expect(load(tmpDir)).rejects.toThrowError(HomeNotFoundError(path.resolve(tmpDir, 'home')))
-  })
-
   it('loads the app when it has a script with a valid configuration', async () => {
     // Given
     await writeConfig(appConfiguration)
-    await mkdirHome()
 
     const blockConfiguration = `
       name = "my-script"
@@ -287,7 +274,6 @@ name = "my_app"
   it('loads the app with several scripts that have valid configurations', async () => {
     // Given
     await writeConfig(appConfiguration)
-    await mkdirHome()
     let blockConfiguration = `
       name = "my-script-1"
       `