+++ packages/app/src/framework/ui-extensions/checkout/react.ts
@@ -1,0 +1,1 @@
+export * from '@shopify/checkout-ui-extensions-react'