+++ packages/app/src/framework/ui-extensions/post-purchase/react.ts
@@ -1,0 +1,1 @@
+export * from '@shopify/post-purchase-ui-extensions-react'