+++ packages/cli-kit/src/output.ts
@@ -7,6 +7,10 @@ enum ContentTokenType {
   Command,
   Path,
   Link,
+  Yellow,
+  Cyan,
+  Magenta,
+  Green,
 }
 
 interface ContentMetadata {
@@ -35,6 +39,18 @@ export const token = {
   link: (value: string, link: string) => {
     return new ContentToken(value, {link}, ContentTokenType.Link)
   },
+  cyan: (value: string) => {
+    return new ContentToken(value, {}, ContentTokenType.Cyan)
+  },
+  yellow: (value: string) => {
+    return new ContentToken(value, {}, ContentTokenType.Yellow)
+  },
+  magenta: (value: string) => {
+    return new ContentToken(value, {}, ContentTokenType.Magenta)
+  },
+  green: (value: string) => {
+    return new ContentToken(value, {}, ContentTokenType.Green)
+  },
 }
 
 // output.content`Something ${output.token.command(Something)}`
@@ -46,7 +62,7 @@ class TokenizedString {
   }
 }
 
-type Message = string | TokenizedString
+export type Message = string | TokenizedString
 
 export function content(strings: TemplateStringsArray, ...keys: (ContentToken | string)[]): TokenizedString {
   let output = ``
@@ -69,6 +85,18 @@ export function content(strings: TemplateStringsArray, ...keys: (ContentToken | 
           break
         case ContentTokenType.Link:
           output += terminalLink(enumToken.value, enumToken.metadata.link ?? '')
+          break
+        case ContentTokenType.Yellow:
+          output += colors.yellow(enumToken.value)
+          break
+        case ContentTokenType.Cyan:
+          output += colors.cyan(enumToken.value)
+          break
+        case ContentTokenType.Magenta:
+          output += colors.magenta(enumToken.value)
+          break
+        case ContentTokenType.Green:
+          output += colors.green(enumToken.value)
           break
       }
     }