+++ packages/create-app/src/commands/init.ts
@@ -16,25 +16,20 @@ export default class Init extends Command {
       parse: (input, _) => Promise.resolve(path.resolve(input)),
       hidden: false,
     }),
+    template: Flags.string({
+      description: 'The template for app home. Eg, --template https://github.com/Shopify/shopify-app-php',
+      env: 'SHOPIFY_FLAG_TEMPLATE',
+    }),
     'dependency-manager': Flags.string({
       char: 'd',
       env: 'SHOPIFY_FLAG_DEPENDENCY_MANAGER',
       hidden: false,
       options: ['npm', 'yarn', 'pnpm'],
     }),
-    'shopify-cli-version': Flags.string({
-      char: 's',
-      env: 'SHOPIFY_FLAG_SHOPIFY_CLI_VERSION',
-      hidden: true,
-    }),
-    'shopify-cli-kit-version': Flags.string({
-      char: 's',
-      env: 'SHOPIFY_FLAG_SHOPIFY_CLI_KIT_VERSION',
-      hidden: true,
-    }),
-    'shopify-app-version': Flags.string({
-      char: 'a',
-      env: 'SHOPIFY_FLAG_SHOPIFY_APP_VERSION',
+    local: Flags.boolean({
+      char: 'l',
+      env: 'SHOPIFY_FLAG_LOCAL',
+      default: false,
       hidden: true,
     }),
   }
@@ -44,13 +39,13 @@ export default class Init extends Command {
     const directory = flags.path ? path.resolve(flags.path) : process.cwd()
     const promptAnswers = await initPrompt({
       name: flags.name,
+      template: flags.template,
     })
     await initService({
       name: promptAnswers.name,
       dependencyManager: flags['dependency-manager'],
-      shopifyCliVersion: flags['shopify-cli-version'],
-      shopifyAppVersion: flags['shopify-app-version'],
-      shopifyCliKitVersion: flags['shopify-cli-kit-version'],
+      template: promptAnswers.template,
+      local: flags.local,
       directory,
     })
   }