+++ packages/create-app/src/constants.ts
@@ -1,0 +1,11 @@
+export const configurationFileNames = {
+  home: 'shopify.home.toml',
+  homeTemplate: 'shopify.home.template.toml',
+}
+
+export const blocks = {
+  home: {
+    directoryName: 'home',
+    configurationName: configurationFileNames.home,
+  },
+}